# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LIFELINK emergency response, dispatch, routing, hospital readiness, and demo simulation API.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a LIFELINK account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountUser"
        "400":
          description: Invalid account data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Sign in to LIFELINK
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/demo-login:
    post:
      operationId: demoLogin
      tags: [auth]
      summary: Sign in to the local operator simulation account
      responses:
        "200":
          description: Local demo operator session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountUser"
        "404":
          description: Demo access is disabled or not seeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current session account
      responses:
        "200":
          description: Current account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountUser"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    AccountUser:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [USER, DRIVER, OPERATOR] }
        phone: { type: ["string", "null"] }
      required: [id, name, email, role, phone]
    RegisterInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
        role: { type: string, enum: [USER, DRIVER] }
        phone: { type: ["string", "null"] }
      required: [name, email, password, role]
    LoginInput:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
      required: [email, password]
...